fix: Move ContainerRegistryRowSchema, ContainerRegistryRowModels under `common` package
