+++ tests/manager/models/test_container_registry_nodes.py
@@ -4,8 +4,8 @@ import pytest
 from graphene import Schema
 from graphene.test import Client
 
+from ai.backend.common.container_registry import ContainerRegistryType
 from ai.backend.manager.defs import PASSWORD_PLACEHOLDER
-from ai.backend.manager.models.container_registry import ContainerRegistryType
 from ai.backend.manager.models.gql import GraphQueryContext, Mutations, Queries
 from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
 from ai.backend.manager.server import database_ctx
